Removed domain-specific messages
Removed SequencerReject and ApplicationDisocvery messages because they are domain-specific and/or too tightly bind implementations.
ApplicationDiscovery is better handled by Metrics announcing when applications turn on/off. Putting it as a first class feature removes determinism in some failover scenarios.
SequencerReject adds a lot of complexity for very little gain in most implementations. For complex entity graphs there could be a use but this should be more narrowly tailored as opposed to being common reject scenario for all non-order messages.