Primary/backup MoldCommandPublishers
Updated the MoldCommandPublisher so it will work better when being promoted from backup to primary.
The publisher will now watch the event stream to check if the events are ahead of the commands being sent. This is what you would expect for a publisher that is operating in backup mode and is non-deterministic from external stimuli (i.e., a FIX port or a telnet command).